chore(implementFeedback): change error response and validate cordinates
 - rename routes
 - change error response to be more dynamic
 - validate cordinates
 - fix bug on updating comment and location

  [delivers #162482661]
