+++ src/controllers/record.js
@@ -1,5 +1,4 @@
 /* eslint-disable max-len */
-import validator from 'validator';
 import Record from '../models/record';
 
 const RecordController = {
@@ -11,13 +10,6 @@ const RecordController = {
    */
   createRecord(req, res) {
     const record = Record.createRecord(req.body);
-    const longLat = req.body.location.trim();
-    if (!validator.isLatLong(longLat)) {
-      return res.status(400).send({
-        status: 400,
-        message: 'Invalid location, please ensure you have valid cordinates. e.g 0.00000,0.00000',
-      });
-    }
     return res.status(201).send({
       status: 201,
       data: [record],
